+++ tests/metrics/bug_139/test_03.m
@@ -1,0 +1,9 @@
+function r = test_03(a, b, c, d)
+
+    if (a | b) & (c | d)
+        r = 'a';
+    else
+        r = 'b';
+    end
+
+end