+++ tests/parser/bug_139/test_02.m
@@ -1,0 +1,9 @@
+function r = test_02(a, b, c, d)
+
+    if [a | b, b & c] == [b | c, c & d]
+        r = 'a';
+    else
+        r = 'b';
+    end
+
+end