+++ tests/parser/overloading_semantics/Potato.m
@@ -1,0 +1,24 @@
+% (c) Copyright 2020 Zenuity AB
+
+classdef Potato
+
+    properties
+        value (1, 1) logical
+    end
+
+    methods
+        function obj = Potato(val)
+            obj.value = val;
+        end
+
+        function result = logical(obj)
+            result = obj.value;
+        end
+
+        function result = and(lhs, rhs)
+            disp("potato");
+            result = lhs.value & rhs.value;
+        end
+    end
+
+end