+++ tests/parser/overloading_semantics/test_01.m
@@ -1,0 +1,11 @@
+% (c) Copyright 2020 Zenuity AB
+
+a = Potato(true);
+b = Potato(false);
+c = Potato(true);
+
+if a & b & c
+    disp("yes");
+else
+    disp("no");
+end